Chatbot/requirements.txt
You will need:

google-generativeai
requests
numexpr